フィールドを複数持てるように修正 ステージスクリプトの書式を変更 ActorManagerでフィールドの衝突判定を引数を取らずに行なうように修正 ActorTag::Fieldを追加 堀

+++ FarbeFahrt/FarbeFahrt/Root/Actor/Gimmick/Gimmick.cpp
@@ -29,7 +29,7 @@ void Gimmick::frameReset()
 //ê”îCé“Å@Ç»ÇËÇΩÇÒ
 void Gimmick::animation()
 {
-	m_maxframe = MV1GetFrameNum(Singleton<HandleList>::Instance().getHandle(m_name));
+	m_maxframe = (float)MV1GetFrameNum(Singleton<HandleList>::Instance().getHandle(m_name));
 	if (isAnimate)
 	{
 		if (frame <= m_maxframe * m_frameSpeed)
